build: avoid a new -Wformat-induced warning
* src/sort.c (mark_key): Add a cast-to-int of a printf field width,
to placate -Wformat.

+++ src/sort.c
@@ -2122,7 +2122,7 @@ count_tabs (char const *text, const size_t len)
 static void
 mark_key (size_t offset, size_t width)
 {
-  printf ("%*s", offset, "");
+  printf ("%*s", (int) offset, "");
 
   if (!width)
     printf (_("^ no match for key\n"));
